+++ cypress/fixtures/createRichTextFakeSdk.tsx
@@ -117,10 +117,14 @@ export function createRichTextFakeSdk(props?: RichTextFakeSdkProps): FieldAppSDK
 
           return store.get('Entry', entryId);
         },
-        getMany: async (query) => {
-          const items: Entry[] = [entries.published, entries.changed, entries.empty];
+        getMany: async ({ query }) => {
+          const items: Entry[] =
+            !query || query.content_type === 'exampleCT'
+              ? [entries.published, entries.changed, entries.empty]
+              : [];
+
           return Promise.resolve({
-            items: !query || query.content_type === 'exampleCT' ? items : [],
+            items,
             total: items.length,
             skip: 0,
             limit: 100,
@@ -136,11 +140,10 @@ export function createRichTextFakeSdk(props?: RichTextFakeSdkProps): FieldAppSDK
 
           return store.get('Asset', assetId);
         },
-        getMany: async (query) => {
-          const items: Asset[] = [
-            assets.published as unknown as Asset,
-            assets.changed as unknown as Asset,
-          ];
+        getMany: async ({ query }) => {
+          const items: Asset[] = query
+            ? [assets.published as unknown as Asset, assets.changed as unknown as Asset]
+            : [];
           return Promise.resolve({
             items: query ? items : [],
             total: items.length,