+++ packages/default-field-editors/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.17](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.16...@contentful/default-field-editors@1.8.17) (2025-09-30)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+## [1.8.16](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.15...@contentful/default-field-editors@1.8.16) (2025-09-30)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+## [1.8.15](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.14...@contentful/default-field-editors@1.8.15) (2025-09-29)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+## [1.8.14](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.13...@contentful/default-field-editors@1.8.14) (2025-09-29)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
+## [1.8.13](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.12...@contentful/default-field-editors@1.8.13) (2025-09-24)
+
+**Note:** Version bump only for package @contentful/default-field-editors
+
 ## [1.8.12](https://github.com/contentful/field-editors/compare/@contentful/default-field-editors@1.8.11...@contentful/default-field-editors@1.8.12) (2025-09-24)
 
 **Note:** Version bump only for package @contentful/default-field-editors