+++ packages/number/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.13](https://github.com/contentful/field-editors/compare/@contentful/field-editor-number@1.5.12...@contentful/field-editor-number@1.5.13) (2025-09-30)
+
+**Note:** Version bump only for package @contentful/field-editor-number
+
+## [1.5.12](https://github.com/contentful/field-editors/compare/@contentful/field-editor-number@1.5.11...@contentful/field-editor-number@1.5.12) (2025-09-29)
+
+**Note:** Version bump only for package @contentful/field-editor-number
+
+## [1.5.11](https://github.com/contentful/field-editors/compare/@contentful/field-editor-number@1.5.10...@contentful/field-editor-number@1.5.11) (2025-09-24)
+
+**Note:** Version bump only for package @contentful/field-editor-number
+
 ## [1.5.10](https://github.com/contentful/field-editors/compare/@contentful/field-editor-number@1.5.9...@contentful/field-editor-number@1.5.10) (2025-09-24)
 
 **Note:** Version bump only for package @contentful/field-editor-number