+++ packages/rating/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.14](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.6.13...@contentful/field-editor-rating@1.6.14) (2025-09-30)
+
+**Note:** Version bump only for package @contentful/field-editor-rating
+
+## [1.6.13](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.6.12...@contentful/field-editor-rating@1.6.13) (2025-09-30)
+
+### Bug Fixes
+
+- improve rating field usability with trackpad [ZEND-6961] ([#1945](https://github.com/contentful/field-editors/issues/1945)) ([1bedee5](https://github.com/contentful/field-editors/commit/1bedee5cb434c0fd32af097235fdaca7078d9f22))
+
+## [1.6.12](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.6.11...@contentful/field-editor-rating@1.6.12) (2025-09-29)
+
+**Note:** Version bump only for package @contentful/field-editor-rating
+
+## [1.6.11](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.6.10...@contentful/field-editor-rating@1.6.11) (2025-09-24)
+
+**Note:** Version bump only for package @contentful/field-editor-rating
+
 ## [1.6.10](https://github.com/contentful/field-editors/compare/@contentful/field-editor-rating@1.6.9...@contentful/field-editor-rating@1.6.10) (2025-09-24)
 
 **Note:** Version bump only for package @contentful/field-editor-rating